+++ test/fixtures/symbols.rb
@@ -17,3 +17,5 @@
 %I[foo]
 %
 %I[foo] # comment
+%
+%I{foo[]}